refactor: update CODE_AGENT import path and expose all configuration from `settings.py`
