+++ src/os_assistant/configs/__init__.py
@@ -6,8 +6,3 @@ import yaml
 CONFIG_PATH = os.path.join(os.path.dirname(__file__), "config.yaml")
 with open(CONFIG_PATH, encoding="utf-8") as config_file:
     CONFIG = yaml.safe_load(config_file)
-
-# Expose specific sections for easier access
-CODE_AGENT = CONFIG.get("CODE_AGENT", {})
-
-# This file makes the configs directory a Python module.
